32-bit vclock counter + FDE support close
Moved to using a 32-bit clock counter, changed the usage of the indices
to be incremented right before assignment by the math functions so we're
always up to date, and am adding support for full testing with CDE and
FDE response with 32-bit numbers

+++ MathCalculations.h
@@ -79,7 +79,7 @@ extern char r;
 extern double phase_correction_factor;
 extern short max_samp;
 
-extern volatile short recbuf_start_clock; // virtual clock counter for first sample in recording buffer
+extern volatile int recbuf_start_clock; // virtual clock counter for first sample in recording buffer
